(svn r18011) -Feature(ette): [NewGRF] CB 36 for roadvehicle property 09 'running cost factor'.

+++ src/train_cmd.cpp
@@ -4438,7 +4438,7 @@ Money Train::GetRunningCost() const
 		const RailVehicleInfo *rvi = RailVehInfo(v->engine_type);
 		if (rvi->running_cost_class == INVALID_PRICE) continue;
 
-		byte cost_factor = GetVehicleProperty(v, PROP_TRAIN_RUNNING_COST_FACTOR, rvi->running_cost);
+		uint cost_factor = GetVehicleProperty(v, PROP_TRAIN_RUNNING_COST_FACTOR, rvi->running_cost);
 		if (cost_factor == 0) continue;
 
 		/* Halve running cost for multiheaded parts */
